simple_music_explorer_api/authapp/serializers.py
from djoser.serializers import UserSerializer, UserCreateSerializer
from rest_framework import serializers

from .models import User


class UserCreateSerializer(UserCreateSerializer):
    class Meta(UserCreateSerializer.Meta):
        model = User
        fields = ('id', 'username', 'email', 'password')
